# docs/rksv/sign-at.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "fiskaly SIGN AT",
...
    "version": "1.2.6",
    "contact": {
      "name": "fiskaly GmbH",
      "url": "https://fiskaly.com",
      "email": "info@fiskaly.com"
    },
    "x-logo": {
      "url": "https://www.fiskaly.com/fiskaly_logo.svg",
      "altText": "fiskaly Logo"
    }
  },
  "components": {
    "schemas": {
...
      "CashRegisterId": {
        "type": "string",
        "format": "uuid",
        "pattern": "[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}",
        "example": "1c81cb86-c2e8-4074-afc3-a0601b2bf063",
        "description": "Identifies a Cash Register."
      },
      "SignatureCreationUnitId": {
        "type": "string",
        "format": "uuid",
        "pattern": "[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}",
        "example": "1c81cb86-c2e8-4074-afc3-a0601b2bf063",
        "description": "Identifies a Signature Creation Unit."
      },
      "ReceiptId": {
        "type": "string",
        "format": "uuid",
        "pattern": "[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}",
        "example": "1c81cb86-c2e8-4074-afc3-a0601b2bf063",
        "description": "Identifies a Receipt."
      },
...
      "Version": {
        "type": "string",
        "default": "1.2.6",
        "example": "1.2.6"
      },
...
      "CreateSignatureCreationUnitRequest": {
        "type": "object",
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/MetadataRequest"
          },
          "legal_entity_id": {
            "$ref": "#/components/schemas/LegalEntityId"
          },
          "legal_entity_name": {
            "$ref": "#/components/schemas/LegalEntityName"
          }
        },
        "required": [
          "legal_entity_id"
        ],
        "additionalProperties": false
      },
      "UpdateSignatureCreationUnitRequest": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "INITIALIZED",
              "DECOMMISSIONED"
            ],
            "example": "INITIALIZED"
          }
        },
        "required": [
          "state"
        ],
        "additionalProperties": false
      },
...
      "AuthenticateFonRequest": {
        "type": "object",
        "properties": {
          "fon_participant_id": {
            "$ref": "#/components/schemas/FonParticipantId"
          },
          "fon_user_id": {
            "$ref": "#/components/schemas/FonUserId"
          },
          "fon_user_pin": {
            "$ref": "#/components/schemas/FonUserPin"
          }
        },
        "required": [
          "fon_participant_id",
          "fon_user_id",
          "fon_user_pin"
        ],
        "additionalProperties": false
      },
...
      "AuthenticateFonResponse": {
        "type": "object",
        "description": "Successful authentication.",
        "properties": {
          "fon_participant_id": {
            "$ref": "#/components/schemas/FonParticipantId"
          },
          "fon_user_id": {
            "$ref": "#/components/schemas/FonUserId"
          },
          "authentication_status": {
            "type": "string",
            "enum": [
              "AUTHENTICATED"
            ],
            "default": "AUTHENTICATED"
          },
          "time_authentication": {
            "$ref": "#/components/schemas/Timestamp"
          }
        },
        "required": [
          "fon_participant_id",
          "fon_user_id",
          "authentication_status",
          "time_authentication"
        ],
        "additionalProperties": false
      },
...
      "CreateCashRegisterRequest": {
        "type": "object",
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/MetadataRequest"
          },
          "description": {
            "$ref": "#/components/schemas/CashRegisterDescription"
          }
        },
        "additionalProperties": false
      },
      "UpdateCashRegisterRequest": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "REGISTERED",
              "INITIALIZED",
              "DECOMMISSIONED",
              "DEFECTIVE",
              "OUTAGE"
            ],
            "example": "INITIALIZED"
          }
        },
        "required": [
          "state"
        ],
        "additionalProperties": false
      },
...
      "CreateSignatureCreationUnitResponse": {
        "type": "object",
        "description": "Returns the created Signature Creation Unit resource.",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/SignatureCreationUnitId"
          },
          "_type": {
            "type": "string",
            "enum": [
              "SIGNATURE_CREATION_UNIT"
            ],
            "default": "SIGNATURE_CREATION_UNIT"
          },
          "_env": {
            "$ref": "#/components/schemas/Environment"
          },
          "_version": {
            "$ref": "#/components/schemas/Version"
          },
          "state": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING"
            ]
          },
          "legal_entity_id": {
            "$ref": "#/components/schemas/LegalEntityId"
          },
          "legal_entity_name": {
            "$ref": "#/components/schemas/LegalEntityName"
          },
          "certificate_serial_number": {
            "$ref": "#/components/schemas/SignatureCreationUnitCertificateSerialNumber"
          },
          "time_pending": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "time_creation": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataRequest"
          }
        },
        "required": [
          "_id",
          "_type",
          "_env",
          "_version",
          "state",
          "legal_entity_id",
          "time_pending"
        ],
        "additionalProperties": false
      },
      "SignatureCreationUnitResponse": {
        "type": "object",
        "description": "Returns the Signature Creation Unit resource.",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/SignatureCreationUnitId"
          },
          "_type": {
            "type": "string",
            "enum": [
              "SIGNATURE_CREATION_UNIT"
            ],
            "default": "SIGNATURE_CREATION_UNIT"
          },
          "_env": {
            "$ref": "#/components/schemas/Environment"
          },
          "_version": {
            "$ref": "#/components/schemas/Version"
          },
          "state": {
            "type": "string",
            "enum": [
              "PENDING",
              "CREATED",
              "INITIALIZED",
              "DECOMMISSIONED",
              "OUTAGE",
              "DEFECTIVE"
            ]
          },
          "legal_entity_id": {
            "$ref": "#/components/schemas/LegalEntityId"
          },
          "legal_entity_name": {
            "$ref": "#/components/schemas/LegalEntityName"
          },
          "certificate_serial_number": {
            "$ref": "#/components/schemas/SignatureCreationUnitCertificateSerialNumber"
          },
          "time_pending": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "time_creation": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataRequest"
          },
          "time_initialization": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "time_decommission": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "time_outage": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "time_defect": {
            "$ref": "#/components/schemas/Timestamp"
          }
        },
        "required": [
          "_id",
          "_type",
          "_env",
          "_version",
          "state",
          "legal_entity_id",
          "time_pending"
        ],
        "additionalProperties": false
      },
      "CreateCashRegisterResponse": {
        "type": "object",
        "description": "Returns the created Cash Register resource.",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/CashRegisterId"
          },
          "_type": {
            "type": "string",
            "enum": [
              "CASH_REGISTER"
            ],
            "default": "CASH_REGISTER"
          },
          "_env": {
            "$ref": "#/components/schemas/Environment"
          },
          "_version": {
            "$ref": "#/components/schemas/Version"
          },
          "state": {
            "type": "string",
            "enum": [
              "CREATED"
            ]
          },
          "serial_number": {
            "$ref": "#/components/schemas/CashRegisterSerialNumber"
          },
          "turnover_counter": {
            "$ref": "#/components/schemas/Numeric"
          },
          "description": {
            "$ref": "#/components/schemas/CashRegisterDescription"
          },
          "time_creation": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataRequest"
          }
        },
        "required": [
          "_id",
          "_type",
          "_env",
          "_version",
          "state",
          "serial_number",
          "time_creation"
        ],
        "additionalProperties": false
      },
...
      "CashRegisterResponse": {
        "type": "object",
        "description": "Returns the updated Cash Register resource",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/CashRegisterId"
          },
          "_type": {
            "type": "string",
            "enum": [
              "CASH_REGISTER"
            ],
            "default": "CASH_REGISTER"
          },
          "_env": {
            "$ref": "#/components/schemas/Environment"
          },
          "_version": {
            "$ref": "#/components/schemas/Version"
          },
          "state": {
            "type": "string",
            "enum": [
              "CREATED",
              "REGISTERED",
              "INITIALIZED",
              "DECOMMISSIONED",
              "OUTAGE",
              "DEFECTIVE"
            ]
          },
          "serial_number": {
            "$ref": "#/components/schemas/CashRegisterSerialNumber"
          },
          "turnover_counter": {
            "$ref": "#/components/schemas/Numeric"
          },
          "description": {
            "$ref": "#/components/schemas/CashRegisterDescription"
          },
          "time_creation": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataRequest"
          },
          "time_registration": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "time_initialization": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "initialization_receipt_id": {
            "$ref": "#/components/schemas/ReceiptId"
          },
          "time_decommission": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "decommission_receipt_id": {
            "$ref": "#/components/schemas/ReceiptId"
          },
          "time_outage": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "time_defect": {
            "$ref": "#/components/schemas/Timestamp"
          }
        },
        "required": [
          "_id",
          "_type",
          "_env",
          "_version",
          "state",
          "serial_number",
          "time_creation"
        ],
        "additionalProperties": false
      },
...
      "FonReceiptValidation": {
        "type": "object",
        "description": "Receipts with a `_type` of `YEARLY_CLOSE`, `INITIALIZATION` or `DECOMMISSION` need to be validated using FON. fiskaly SIGN AT does this automatically. This property reports the status of this validation process.",
        "properties": {
          "validation_result": {
            "$ref": "#/components/schemas/FonReceiptValidationResult"
          },
          "time_validation": {
            "$ref": "#/components/schemas/Timestamp"
          }
        },
        "required": [
          "validation_result",
          "time_validation"
        ],
        "additionalProperties": false
      },
...
      "NotFoundErrorResponse": {
        "type": "object",
        "description": "Resource not found.",
        "properties": {
          "status_code": {
            "type": "integer",
            "description": "The status code returned by the API.",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Textual description of the meaning of the status code.",
            "example": "Bad Request"
          },
          "code": {
            "type": "string",
            "description": "Code of the error.",
            "example": "E_BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "description": "Error message further explaining the source of the issue.",
            "example": "SCU does not exist."
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "description": "A JSON Web Token (JWT) used for access control and authorization.\n- Usage format: `Bearer <JWT>`\n- Example HTTP header: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/signature-creation-unit/{signature_creation_unit_id}": {
      "put": {
        "operationId": "createSignatureCreationUnit",
        "summary": "Create a Signature Creation Unit",
        "tags": [
          "Signature Creation Units"
        ],
        "description": "This endpoint creates a Signature Creation Unit. \nThe Signature Creation Unit is identified by a `signature_creation_unit_id`. \nThe `signature_creation_unit_id` must comply with the [UUIDv4](#section/Introduction/UUIDv4) standard and should be generated on your side. \n\nWhen you create a Signature Creation Unit, its `state` is usually set to `CREATED`.\n\nIn some rare cases it may happen that the Signature Creation Unit can not be created immediately.\nThis will be reflected by a transient `PENDING` `state` that will eventually change to `CREATED` at a later point in time.\nSimply retry the request or use the [Retrieve a Signature Creation Unit](#operation/retrieveSignatureCreationUnit) endpoint to regularly check the `state` of the Signature Creation Unit.\n\nUse the [Update a Signature Creation Unit](#operation/updateSignatureCreationUnit) endpoint to transition a Signature Creation Unit to the `state` `INITIALIZED` in order to be able to use it for signing Receipts.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSignatureCreationUnitRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "signature_creation_unit_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SignatureCreationUnitId"
            }
          }
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSignatureCreationUnitResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "retrieveSignatureCreationUnit",
        "summary": "Retrieve a Signature Creation Unit",
        "tags": [
          "Signature Creation Units"
        ],
        "description": "This endpoint returns information about a specific Signature Creation Unit.",
        "parameters": [
          {
            "in": "path",
            "name": "signature_creation_unit_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SignatureCreationUnitId"
            }
          }
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignatureCreationUnitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateSignatureCreationUnit",
        "summary": "Update a Signature Creation Unit",
        "tags": [
          "Signature Creation Units"
        ],
        "description": "This endpoint updates a Signature Creation Unit. \n\nIn order to be able to use a Signature Creation Unit for signing Receipts it needs to be transitioned to the `state` `INITIALIZED`.\n\nOnce a Signature Creation Unit has eventually reached its end-of-life this has to be reflected by setting its `state` to `DECOMMISSIONED`.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSignatureCreationUnitRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "signature_creation_unit_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SignatureCreationUnitId"
            }
          }
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignatureCreationUnitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cash-register/{cash_register_id}": {
      "put": {
        "operationId": "createCashRegister",
        "summary": "Create a Cash Register",
        "tags": [
          "Cash Registers"
        ],
        "description": "This endpoint creates a Cash Register. \nThe Cash Register is identified by a `cash_register_id`. \nThe `cash_register_id` must comply with the [UUIDv4](#section/Introduction/UUIDv4) standard and should be generated on your side. \n\nWhen you create a Cash Register, its `state` is set to `CREATED`.\n\nUse the [Update a Cash Register](#operation/updateCashRegister) endpoint to transition a Cash Register to the `state` `INITIALIZED` in order to be able to use it for signing Receipts.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCashRegisterRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "cash_register_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CashRegisterId"
            }
          }
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCashRegisterResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "retrieveCashRegister",
        "summary": "Retrieve a Cash Register",
        "tags": [
          "Cash Registers"
        ],
        "description": "This endpoint returns information about a specific Cash Register.",
        "parameters": [
          {
            "in": "path",
            "name": "cash_register_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CashRegisterId"
            }
          }
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashRegisterResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateCashRegister",
        "summary": "Update a Cash Register",
        "tags": [
          "Cash Registers"
        ],
        "description": "This endpoint updates a Cash Register. \n\nIn order to be able to use a Cash Register for signing Receipts it needs to be transitioned to the `state` `INITIALIZED`.\n\nOnce a Cash Register has eventually reached its end-of-life this has to be reflected by setting its `state` to `DECOMMISSIONED`.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCashRegisterRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "cash_register_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CashRegisterId"
            }
          }
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashRegisterResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cash-register/{cash_register_id}/receipt/{receipt_id}/validation": {
      "post": {
        "operationId": "validateReceipt",
        "summary": "Validate a Receipt",
        "tags": [
          "Receipts"
        ],
        "description": "This endpoint validates a Receipt.\n\nThis returns a status representing the outcome of the validation against FinanzOnline, as well as a message explaining the validation result.\nThe validation is performed in a sychronous manner, meaning if FinanzOnline is not available the request might time out.",
        "parameters": [
          {
            "in": "path",
            "name": "cash_register_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CashRegisterId"
            }
          },
          {
            "in": "path",
            "name": "receipt_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReceiptId"
            }
          }
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FonReceiptValidation"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fon/auth": {
      "put": {
        "operationId": "authenticateFon",
        "summary": "Authenticate FON",
        "tags": [
          "FON",
          "Authentication"
        ],
        "description": "The fiscalization process requires interactions with [FinanzOnline (FON)](#section/Resources/FinanzOnline-(FON)).\nThis includes registering Cash Registers and Signature Creation Units, validating Receipts, etc.\nThe fiskaly SIGN AT automates all those steps.\n\nOur system communicates with [FinanzOnline (FON)](https://finanzonline.bmf.gv.at/fon/) in the name of the taxpayer. \nThe taxpayer must create a [FinanzOnline Cash Register web service user (\"*Registrierkassen-Webservice-Benutzer*\", pages 60-63)](https://finanzonline.bmf.gv.at/eLearning/BMF_Handbuch_Registrierkassen.pdf). \nThen you must use those credentials to authenticate the taxpayer with FON through this endpoint.\n\n**Note**:\nThe credentials of this Cash Register Web Service User must be provided **only once** for authenticating with FON.\nThe Cash Register web service user remains authenticated with FON indefinitely.\n\n**Note**: \nYou must authenticate with FON before transitioning Signature Creation Units and Cash Registers to `INITIALIZED`. \n\n**Note**:\nFor requests requiring our system to communicate with [FinanzOnline (FON)](https://finanzonline.bmf.gv.at/fon/), \nany downtime or unavailability of the [FinanzOnline (FON)](https://finanzonline.bmf.gv.at/fon/) service may impact such requests.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateFonRequest"
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticateFonResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "retrieveFonStatus",
        "summary": "Retrieve FON Status",
        "tags": [
          "FON"
        ],
        "description": "This endpoint can be used to check the status of the FON authentication.\n\nIf `authentication_status` is not\n`AUTHENTICATED` then it is necessary that the user reauthenticates with FON using the [Authenticate FON](#operation/authenticateFon) endpoint.",
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FonStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configuration": {
      "patch": {
        "operationId": "updateConfiguration",
        "summary": "Update organization's configuration",
        "tags": [
          "Configurations"
        ],
        "description": "This endpoint updates current Configuration for an Organization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "yearly_receipt_validation_enabled": true,
                  "monthly_receipt_validation_enabled": true
                },
                "properties": {
                  "yearly_receipt_validation_enabled": {
                    "type": "boolean"
                  },
                  "monthly_receipt_validation_enabled": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "retrieveConfiguration",
        "summary": "Retrieve organization's configuration",
        "tags": [
          "Configurations"
        ],
        "description": "This endpoint retrieves the current Configuration for an Organization.",
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "https://rksv.fiskaly.com"
    }
  ],
...
}